Moved core to libs
