+++ src/componentsV1/PlatformHeaderView.tsx
@@ -1,10 +1,10 @@
 import Grid from '@material-ui/core/Grid';
 import React from 'react';
 
-import { MintosPlatform } from '../core/platforms/MintosPlatform';
-import { IBaseResult } from '../core/platforms/models';
-import { TwinoPlatform } from '../core/platforms/TwinoPlatform';
-import { ZonkyPlatform } from '../core/platforms/ZonkyPlatform';
+import { MintosPlatform } from '../libs/core/platforms/MintosPlatform';
+import { IBaseResult } from '../libs/core/platforms/models';
+import { TwinoPlatform } from '../libs/core/platforms/TwinoPlatform';
+import { ZonkyPlatform } from '../libs/core/platforms/ZonkyPlatform';
 
 import MintosSpecificHeaderView from './platformSpecific/MintosSpecificHeaderView';
 import TwinoSpecificHeaderView from './platformSpecific/TwinoSpecificHeaderView';
