+++ src/libs/core/platforms/__tests__/MintosPlatform201911.ts
